+++ contracts/Vault.sol
@@ -13,6 +13,7 @@ import "./rewards/RewardsProtocol.sol";
 import "./riskmgmt/RiskMgmtProtocol.sol";
 import "./exchange/ExchangeProtocol.sol";
 import "./VaultProtocol.sol";
+import "./dependencies/Logger.sol";
 
 /// @title Vault Contract
 /// @author Melonport AG <team@melonport.com>
@@ -62,16 +63,9 @@ contract Vault is DBC, Owned, Shares, VaultProtocol {
     Prospectus public prospectus;
     Modules public module;
     Calculations public atLastPayout;
+    Logger public logger;
 
     // EVENTS
-
-    event Subscribed(address indexed byParticipant, uint256 atTimestamp, uint256 numShares);
-    event Redeemed(address indexed byParticipant, uint256 atTimestamp, uint256 numShares);
-    event PortfolioContent(uint256 assetHoldings, uint256 assetPrice, uint256 assetDecimals); // Calcualtions
-    event SpendingApproved(address ofToken, address onExchange, uint256 amount); // Managing
-    event RewardsConverted(uint256 atTimestamp, uint256 numSharesConverted, uint256 numunclaimedRewards);
-    event RewardsPayedOut(uint256 atTimestamp, uint256 numSharesPayedOut, uint256 atSharePrice);
-    event CalculationUpdate(uint256 atTimestamp, uint256 managementReward, uint256 performanceReward, uint256 nav, uint256 sharePrice, uint256 totalSupply);
 
     // PRE, POST, INVARIANT CONDITIONS
 
@@ -175,7 +169,7 @@ contract Vault is DBC, Owned, Shares, VaultProtocol {
               assetPrice = Price.getPrice(ofAsset); // Asset price given quoted to melonAsset (and 'quoteAsset') price
             }
             gav = gav.add(assetHoldings.mul(assetPrice).div(10 ** uint(assetDecimals))); // Sum up product of asset holdings of this vault and asset prices
-            PortfolioContent(assetHoldings, assetPrice, assetDecimals);
+            logger.logPortfolioContent(assetHoldings, assetPrice, assetDecimals);
         }
     }
 
@@ -190,8 +184,10 @@ contract Vault is DBC, Owned, Shares, VaultProtocol {
         address ofUniverse,
         address ofParticipation,
         address ofRiskMgmt,
-        address ofRewards
+        address ofRewards,
+        address ofLogger
     ) {
+        logger = Logger(ofLogger);
         owner = ofManager;
         name = withName;
         symbol = withSymbol;
@@ -245,7 +241,7 @@ contract Vault is DBC, Owned, Shares, VaultProtocol {
             assert(offeredValue >= actualValue); // Sanity Check
             assert(AssetProtocol(melonAsset).transferFrom(msg.sender, this, actualValue));  // Transfer value
             createShares(msg.sender, numShares); // Accounting
-            Subscribed(msg.sender, now, numShares);
+            logger.logSubscribed(msg.sender, now, numShares);
         }
     }
 
@@ -261,7 +257,7 @@ contract Vault is DBC, Owned, Shares, VaultProtocol {
         assert(requestedValue <= actualValue); // Sanity Check
         assert(AssetProtocol(melonAsset).transfer(msg.sender, actualValue)); // Transfer value
         annihilateShares(msg.sender, numShares); // Accounting
-        Redeemed(msg.sender, now, numShares);
+        logger.logRedeemed(msg.sender, now, numShares);
     }
 
     /// Pre: Approved spending of all assets with non-empty asset holdings;
@@ -272,7 +268,7 @@ contract Vault is DBC, Owned, Shares, VaultProtocol {
         pre_cond(isPastZero(numShares))
     {
         allocateSlice(numShares);
-        Subscribed(msg.sender, now, numShares);
+        logger.logSubscribed(msg.sender, now, numShares);
     }
 
     /// Pre: Recipient owns shares
@@ -282,7 +278,7 @@ contract Vault is DBC, Owned, Shares, VaultProtocol {
         pre_cond(balancesOfHolderAtLeast(msg.sender, numShares))
     {
         separateSlice(numShares);
-        Redeemed(msg.sender, now, numShares);
+        logger.logRedeemed(msg.sender, now, numShares);
     }
 
     /// Pre: Allocation: Pre-approve spending for all non empty vaultHoldings of Assets, numShares denominated in [base units ]
@@ -422,7 +418,7 @@ contract Vault is DBC, Owned, Shares, VaultProtocol {
         internal
     {
         assert(ofToken.approve(onExchange, amount));
-        SpendingApproved(ofToken, onExchange, amount);
+        logger.logSpendingApproved(ofToken, onExchange, amount);
     }
 
     // NON-CONSTANT METHODS - REWARDS
@@ -457,7 +453,7 @@ contract Vault is DBC, Owned, Shares, VaultProtocol {
           timestamp: now
         });
 
-        RewardsConverted(now, numShares, unclaimedRewards);
-        CalculationUpdate(now, managementReward, performanceReward, nav, sharePrice, totalSupply);
+        logger.logRewardsConverted(now, numShares, unclaimedRewards);
+        logger.logCalculationUpdate(now, managementReward, performanceReward, nav, sharePrice, totalSupply);
     }
 }